PKCS11dotNetV2 fix filenames case
There was a bug https://github.com/svn-all-fast-export/svn2git/pull/21
in the Subversion-to-Git conversion that did not correctly process
case-only renames. There appear to have been some case-only renames in
the history of the Subversion repository. As a consequence of this bug,
some files in the Git repository now have names with the wrong case, and
need to be renamed.

These files:

* SmartCardServices/src/PKCS11dotNetV2/application.cpp
* SmartCardServices/src/PKCS11dotNetV2/cardmoduleservice.cpp
* SmartCardServices/src/PKCS11dotNetV2/digest.cpp
* SmartCardServices/src/PKCS11dotNetV2/log.cpp
* SmartCardServices/src/PKCS11dotNetV2/pkcs11.cpp
* SmartCardServices/src/PKCS11dotNetV2/session.cpp
* SmartCardServices/src/PKCS11dotNetV2/slot.cpp
* SmartCardServices/src/PKCS11dotNetV2/template.cpp
* SmartCardServices/src/PKCS11dotNetV2/timer.cpp

Should be renamed to:

* SmartCardServices/src/PKCS11dotNetV2/Application.cpp
* SmartCardServices/src/PKCS11dotNetV2/CardModuleService.cpp
* SmartCardServices/src/PKCS11dotNetV2/Digest.cpp
* SmartCardServices/src/PKCS11dotNetV2/Log.cpp
* SmartCardServices/src/PKCS11dotNetV2/PKCS11.cpp
* SmartCardServices/src/PKCS11dotNetV2/Session.cpp
* SmartCardServices/src/PKCS11dotNetV2/Slot.cpp
* SmartCardServices/src/PKCS11dotNetV2/Template.cpp
* SmartCardServices/src/PKCS11dotNetV2/Timer.cpp

This problem affects master, but might also affect branches and tags.

Thanks to Ryan Schmidt for the bug report
"Wrong case of filenames in PKCS11dotNetV2"
https://github.com/smartcardservices/smartcardservices/issues/144